Create code
This code creates a Crossword object representing a crossword grid of size x size. The add_word method adds a word to the crossword at the specified position (row, column) and direction with the specified word length. The display method prints the crossword grid.

In this example, the words "PYTHON" and "JAVA" are added to the crossword horizontally and vertically respectively.